% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEAforR_functions.R
\name{load_GSEA_results}
\alias{load_GSEA_results}
\title{Load GSEA results}
\usage{
load_GSEA_results(gsea_results_dir)
}
\arguments{
\item{gsea_results_dir}{GSEA results directory filepath}
}
\value{
Dataframe of GSEA Hallmark Gene Set Analysis results
}
\description{
Loads the GSEA Hallmark Gene Set Analysis results summary files
}
