% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEAforR_functions.R
\name{normalize_and_run_GSEA}
\alias{normalize_and_run_GSEA}
\title{Normalize data and run GSEA}
\usage{
normalize_and_run_GSEA(
  expr_data,
  groups,
  control,
  permute_mode,
  out_name,
  out_dir,
  GSEAforR
)
}
\arguments{
\item{expr_data}{Gene expression dataset (dataframe or TSV). The first column must be named "hgnc_symbol".}

\item{groups}{Experimental group metadata for each sample (dataframe or TSV). The columns must be named "sample_id", "group". Group names may not contain special characters or spaces.}

\item{control}{Name of the control group.}

\item{permute_mode}{Whether to perform GSEA permutations by "phenotype" (random assignment of sample groups) or "gene_set" (randomly select genes).}

\item{out_name}{Name under which to save output files.}

\item{out_dir}{Directory in which to save output files.}

\item{GSEAforR}{Location of the GSEAforR directory.}
}
\value{
No return value: saves to file.
}
\description{
Function to perform count normalization using the DESeq2 package and GSEA analysis using the GSEA command line interface
}
