% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEAforR_functions.R
\name{plot_GSEA_results}
\alias{plot_GSEA_results}
\title{Plot GSEA results}
\usage{
plot_GSEA_results(
  result,
  fdr_q_cutoff,
  control_group,
  experimental_group,
  order = 1,
  save_file = NULL
)
}
\arguments{
\item{result}{GSEA result, loaded using `load_GSEA_results()`}

\item{fdr_q_cutoff}{Desired significance cutoff for FDR Q-value}

\item{control_group}{Name of the control group}

\item{experimental_group}{Name of the experimental group}

\item{order}{Order the pathways from largest to smallest NES value (order=1) or smallest to largest NES value (order=-1). Default value: 1}

\item{save_file}{(Optional) Save the plot to this filepath}
}
\value{
Waterfall plot of GSEA results, with significant results indicated by colour
}
\description{
Creates a waterfall plot of the GSEA Hallmark Gene Set Analysis results summary
}
